+++ app/scripts/fhir/R4/sdc-export.js
@@ -54,7 +54,7 @@ var self = {
     var rtn = [qr];
     for (var i=0, len=lfData.itemList.length; i<len; ++i) {
       var item = lfData.itemList[i];
-      if (item._fhirExt && item._fhirExt[this.fhirExtObsLinkPeriod] && item.value) {
+      if (self._getExtractValue(item) && item.value) {
         var obs = this._commonExport._createObservation(item);
         for (var j=0, jLen=obs.length; j<jLen; j++) {
           // Following