+++ test/protractor/spec/R4/prepop.spec.js
@@ -111,7 +111,7 @@ describe('Form pre-population', function() {
           expect(weightField.getAttribute('value')).toBe('95');
         });
 
-        it('should extract observationLinkPeriod fields that are not top-level', function(done) {
+        it('should extract observationExtract fields that are not top-level', function(done) {
           // Follows previous test on population
           var releaseVersion = serverFHIRNum == '3.0' ? 'STU3' : 'R4';
           var resourcesPromise = browser.executeScript(
@@ -139,14 +139,14 @@ describe('Form pre-population', function() {
           expect(weightField.getAttribute('value')).toBe('96');
         });
 
-        it('should extract observationLinkPeriod fields when multiple codes are present', function (done) {
+        it('should extract observationExtract fields when multiple codes are present', function (done) {
           // Follows previous test on population
           var releaseVersion = serverFHIRNum == '3.0' ? 'STU3' : 'R4';
           var resourcesPromise = browser.executeScript(
             'return LForms.Util.getFormFHIRData("QuestionnaireResponse", "'+
             releaseVersion + '", null, {"extract": true})');
           resourcesPromise.then((resources) => {
-            expect(resources.length).toBe(2); // One QR and one observation
+            expect(resources.length).toBe(3); // One QR and two observations
             var obs = resources[1];
             expect(obs.code.coding.length).toEqual(2);
             expect(obs.resourceType).toBe("Observation");